Adding proper sidebar with routes, still overshadowed by banner
