+++ src/components/Navbar.vue
@@ -4,6 +4,7 @@
       <v-app-bar app elevate-on-scroll v-scroll="floatingNav" id="red-header">
         <v-app-bar-nav-icon
           class="d-flex d-sm-flex d-md-none d-lg-none text-shadow red-header--icon"
+          @click="drawer = !drawer"
         ></v-app-bar-nav-icon>
         <img
           v-if="white"
@@ -34,6 +35,7 @@
       <v-app-bar app elevate-on-scroll v-scroll="floatingNav" id="gray-header">
         <v-app-bar-nav-icon
           class="d-flex d-sm-flex d-md-none d-lg-none text-shadow red-header--icon"
+          @click="drawer = !drawer"
         ></v-app-bar-nav-icon>
         <v-spacer></v-spacer>
         <span style="color:black">
@@ -46,6 +48,7 @@
       <v-app-bar app elevate-on-scroll v-scroll="floatingNav" id="red-header">
         <v-app-bar-nav-icon
           class="d-flex d-sm-flex d-md-none d-lg-none text-shadow red-header--icon"
+          @click="drawer = !drawer"
         ></v-app-bar-nav-icon>
       </v-app-bar>
     </template>
@@ -53,6 +56,7 @@
       <v-app-bar app elevate-on-scroll v-scroll="floatingNav" id="gray-header">
         <v-app-bar-nav-icon
           class="d-flex d-sm-flex d-md-none d-lg-none text-shadow red-header--icon"
+          @click="drawer = !drawer"
         ></v-app-bar-nav-icon>
         <v-spacer></v-spacer>
         <span style="color:black">
@@ -147,6 +151,32 @@
         <v-spacer></v-spacer>
       </v-app-bar>
     </template>
+    <v-navigation-drawer app v-model="drawer" temporary color="#efefef">
+      <v-list-item>
+        <v-list-item-content>
+          <v-list-item-title style="color: black">
+            <center>QR COMER</center>
+          </v-list-item-title>
+        </v-list-item-content>
+      </v-list-item>
+
+      <v-divider></v-divider>
+
+      <v-list dense>
+        <v-list-item v-for="item in items" :key="item.title" link>
+          <a :href="item.link" style="color: transparent">
+            <v-list-item-icon>
+              <v-icon style="color: black" @click="routeTo(link)">{{ item.icon }}</v-icon>
+            </v-list-item-icon>
+          </a>
+          <a :href="item.link" style="color: transparent">
+            <v-list-item-content>
+              <v-list-item-title style="color: black">{{ item.title }}</v-list-item-title>
+            </v-list-item-content>
+          </a>
+        </v-list-item>
+      </v-list>
+    </v-navigation-drawer>
   </div>
 </template>
 
@@ -157,10 +187,42 @@ import { routeTo } from "../services/context";
 export default {
   data() {
     return {
+      drawer: false,
       routeName: this.$router.currentRoute.name,
       white: false,
-      shoppingCNPJ: ""
+      shoppingCNPJ: "",
+      items: [
+        { title: "Home", icon: "mdi-home", link: "/" },
+        {
+          title: "Shopping",
+          icon: "mdi-store",
+          link: "/shopping/:cnpj"
+        },
+        { title: "Minha sacola", icon: "mdi-shopping", link: "/sacola" },
+        { title: "Meu perfil", icon: "mdi-account", link: "/usuario" },
+        { title: "Sair", icon: "mdi-logout", link: "/logout" }
+      ]
     };
+  },
+  computed: {
+    home: function() {
+      return this.$t("Sidebar.home");
+    },
+    listShoppings: function() {
+      return this.$t("SideBar.shopping");
+    },
+
+    myBag: function() {
+      return this.$t("SideBar.sacola");
+    },
+
+    myProfile: function() {
+      return this.$t("SideBar.profile");
+    },
+
+    getOut: function() {
+      return this.$t("SideBar.logout");
+    }
   },
   computed: mapGetters({
     isAuthenticated: "auth/isAuthenticated"
