--- src/views/SidebarTest.vue
@@ -1,12 +1,0 @@
-<script>
-import UserBanner from "../components/Sidebar";
-
-export default {
-  components: {
-    Sidebar
-  },
-  data() {
-    return {};
-  }
-};
-</script>
